# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion des dettes et créances",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Gestion des dettes et récues"
        }
    ],
    "paths": {
...
        "/dettes": {
            "get": {
                "tags": [
                    "Dettes"
                ],
                "summary": "Lister ensemble des dettes",
                "description": "Liste toutes les dettes avec filtre optionnel de statut. Solde (montantRestant=0) ou NonSolde (montantRestant!=0). Retourne les dettes sans articles + client. (Rôle Boutiquier requis)",
                "operationId": "271cccc424ec9cadaf0baf93c11088ac",
                "parameters": [
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut: Solde (montantRestant=0) ou NonSolde (montantRestant!=0)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Solde",
                                "NonSolde"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succes - Liste des dettes ou Pas Dettes si vide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "oneOf": [
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Dette"
                                                    }
                                                },
                                                {
                                                    "type": "null"
                                                }
                                            ]
                                        },
                                        "message": {
                                            "description": "'Liste des dettes' si trouvees, 'Pas Dettes' si vide",
                                            "type": "string",
                                            "example": "Liste des dettes"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dettes"
                ],
                "summary": "Enregistrer une nouvelle dette",
                "description": "Crée une nouvelle dette avec articles et paiement optionnel. Met à jour le stock des articles. (Rôle Boutiquier requis)",
                "operationId": "6443282c0740c620f6f5b0e148b3eff8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreDetteRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dette enregistree avec succees",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Dette"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dette enregistree avec succees"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "411": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 411
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dettes/{id}": {
            "get": {
                "tags": [
                    "Dettes"
                ],
                "summary": "Lister une dette a partir de l'id",
                "description": "Récupère les détails d'une dette spécifique avec le client (sans articles ni paiements)",
                "operationId": "1ebd7f767beaec214f1deff4aa421013",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la dette à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dette trouvee",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Dette"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dette trouvee"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "411": {
                        "description": "Objet non trouve",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 411
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Objet non trouve"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dettes/{id}/articles": {
            "post": {
                "tags": [
                    "Dettes"
                ],
                "summary": "Lister les articles d'une dette",
                "description": "Récupère une dette avec ses articles détaillés (Rôles: Boutiquier et Client)",
                "operationId": "cdc92b9044db92b468c2290e8bb20d73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la dette",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dette avec articles trouvee",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Dette"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dette avec articles trouvee"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "411": {
                        "description": "Objet non trouve",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 411
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Objet non trouve"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dettes/{id}/paiements": {
            "get": {
                "tags": [
                    "Dettes"
                ],
                "summary": "Lister les paiements d'une dette",
                "description": "Récupère une dette avec ses paiements (Rôles: Boutiquier et Client)",
                "operationId": "1a98cf4991599a4e6ce8be0b81f37b3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la dette",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dette avec paiements trouvee",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Dette"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dette avec paiements trouvee"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "411": {
                        "description": "Objet non trouve",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 411
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Objet non trouve"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}